+++ licensecheck/license_matrix.py
@@ -166,8 +166,8 @@ L_GPL_3_INCOMPATIBLE = [L.GPL_2]
 def depCompatWMyLice(
 	myLicense: L,
 	depLice: list[L],
-	ignoreLicenses: list[L] = None,
-	failLicenses: list[L] = None,
+	ignoreLicenses: list[L] | None = None,
+	failLicenses: list[L] | None = None,
 ) -> bool:
 	"""Identify if the end user license is compatible with the dependency license(s).